# n0stack.swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "title": "n0stack",
    "contact": {
      "name": "n0stack",
      "url": "https://github.com/n0stack/n0stack"
    },
    "version": "version not set"
  },
  "paths": {
    "/api/v0/block_storage": {
      "get": {
        "tags": [
          "BlockStorageService"
        ],
        "operationId": "ListBlockStorages",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v0ListBlockStoragesResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "BlockStorageService"
        ],
        "operationId": "CreateBlockStorage",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v0CreateBlockStorageRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v0BlockStorage"
            }
          }
        }
      }
    },
    "/api/v0/block_storage/{name}": {
      "get": {
        "tags": [
          "BlockStorageService"
        ],
        "operationId": "GetBlockStorage",
        "parameters": [
          {
            "type": "string",
            "name": "name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v0BlockStorage"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "BlockStorageService"
        ],
        "operationId": "DeleteBlockStorage",
        "parameters": [
          {
            "type": "string",
            "name": "name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v0BlockStorage"
            }
          }
        }
      }
    },
    "/api/v0/block_storage/{name}:download": {
      "get": {
        "tags": [
          "BlockStorageService"
        ],
        "operationId": "DownloadBlockStorage",
        "parameters": [
          {
            "type": "string",
            "name": "name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v0DownloadBlockStorageResponse"
            }
          }
        }
      }
    },
    "/api/v0/block_storage/{name}:purge": {
      "post": {
        "tags": [
          "BlockStorageService"
        ],
        "operationId": "PurgeBlockStorage",
        "parameters": [
          {
            "type": "string",
            "name": "name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {}
          }
        }
      }
    },
    "/api/v0/block_storage/{name}:undelete": {
      "post": {
        "tags": [
          "BlockStorageService"
        ],
        "operationId": "UndeleteBlockStorage",
        "parameters": [
          {
            "type": "string",
            "name": "name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v0BlockStorage"
            }
          }
        }
      }
    },
    "/api/v0/block_storage:copy": {
      "post": {
        "tags": [
          "BlockStorageService"
        ],
        "operationId": "CopyBlockStorage",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v0CopyBlockStorageRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v0BlockStorage"
            }
          }
        }
      }
    },
    "/api/v0/block_storage:fetch": {
      "post": {
        "tags": [
          "BlockStorageService"
        ],
        "operationId": "FetchBlockStorage",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v0FetchBlockStorageRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v0BlockStorage"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "BlockStorageBlockStorageState": {
      "type": "string",
      "title": "- PENDING: working API",
      "default": "BLOCK_STORAGE_UNSPECIFIED",
      "enum": [
        "BLOCK_STORAGE_UNSPECIFIED",
        "PENDING",
        "AVAILABLE",
        "IN_USE",
        "PROTECTED",
        "DELETED"
      ]
    },
...
    "v0BlockStorage": {
      "type": "object",
      "properties": {
        "annotations": {
          "description": "Annotations can store metadata used by the system for control.\nIn particular, implementation-dependent fields that can not be set as protobuf fields are targeted.\nThe control specified by n0stack may delete metadata specified by the user.",
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "labels": {
          "description": "Labels stores user-defined metadata.\nThe n0stack system must not rewrite this value.",
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "limit_bytes": {
          "type": "string",
          "format": "uint64"
        },
        "name": {
          "description": "Name is a unique field.",
          "type": "string"
        },
        "node_name": {
          "type": "string"
        },
        "request_bytes": {
          "type": "string",
          "format": "uint64"
        },
        "state": {
          "$ref": "#/definitions/BlockStorageBlockStorageState"
        },
        "storage_name": {
          "type": "string"
        }
      }
    },
...
    "v0CopyBlockStorageRequest": {
      "type": "object",
      "properties": {
        "annotations": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "labels": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "limit_bytes": {
          "type": "string",
          "format": "uint64"
        },
        "name": {
          "type": "string"
        },
        "request_bytes": {
          "type": "string",
          "format": "uint64"
        },
        "source_block_storage": {
          "type": "string"
        }
      }
    },
    "v0CreateBlockStorageRequest": {
      "type": "object",
      "properties": {
        "annotations": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "labels": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "limit_bytes": {
          "type": "string",
          "format": "uint64"
        },
        "name": {
          "type": "string"
        },
        "request_bytes": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
...
    "v0DownloadBlockStorageResponse": {
      "type": "object",
      "properties": {
        "download_url": {
          "type": "string"
        }
      }
    },
    "v0FetchBlockStorageRequest": {
      "type": "object",
      "properties": {
        "annotations": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "labels": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "limit_bytes": {
          "type": "string",
          "format": "uint64"
        },
        "name": {
          "type": "string"
        },
        "request_bytes": {
          "type": "string",
          "format": "uint64"
        },
        "source_url": {
          "type": "string"
        }
      }
    },
    "v0ListBlockStoragesResponse": {
      "type": "object",
      "properties": {
        "block_storages": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v0BlockStorage"
          }
        }
      }
    },
...
  }
}